backend/auth-service/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-auth==2.23.0
requests>=2.31.0